+++ packages/puppeteer/src/window.ts
@@ -1,0 +1,5 @@
+declare global {
+    interface Window {
+        runZeroReport: () => Promise<void>;
+    }
+}